+++ lib/resque/data_store.rb
@@ -76,9 +76,10 @@ module Resque
       @redis.inspect
     end
 
-    # Force a reconnect to Redis.
+    # Force a reconnect to Redis without closing the connection in the parent
+    # process after a fork.
     def reconnect
-      @redis._client.reconnect
+      @redis._client.connect
     end
 
     # Returns an array of all known Resque keys in Redis. Redis' KEYS operation