+++ HISTORY.md
@@ -1,5 +1,8 @@
 ## Unreleased
 
+### Fixed
+* Fix closing parent Redis connection in child process after fork
+
 ### Added
 
 *